add asset

+++ ultralytics/data/dataset.py
@@ -469,7 +469,7 @@ class PotsamDataset(BaseDataset):
     def split_mask(self, mask):
         nc, colors = self.data['nc'], self.data['colors']
         h, w, c = mask.shape
-        results = np.zeros((h, w, nc), dtype=np.float_)
+        results = np.zeros((h, w, nc), dtype=np.float32)
         _,mask_b = cv2.threshold(mask[:,:,0], 125, 255, cv2.THRESH_BINARY)
         _,mask_g = cv2.threshold(mask[:, :, 1], 125, 255, cv2.THRESH_BINARY)
         _,mask_r = cv2.threshold(mask[:, :, 2], 125, 255, cv2.THRESH_BINARY)
